Port to old GCC

+++ trunk/app/warcfilter.c
@@ -163,8 +163,8 @@ int main (int argc, const char ** argv)
 
       /* dump WRecord */
 
-      fprintf (stdout, "%-10llu ",  WRecord_getOffset (r));
-      fprintf (stdout, "%-10llu ",  WRecord_getCompressedSize (r));
+      fprintf (stdout, "%-10llu ", (unsigned long long) WRecord_getOffset (r));
+      fprintf (stdout, "%-10llu ", (unsigned long long) WRecord_getCompressedSize (r));
       fprintf (stdout, "%-10s ",   WRecord_getWarcId        (r));
       fprintf (stdout, "%-10u ",   WRecord_getDataLength    (r));
       fprintf (stdout, "%-15u ",   WRecord_getRecordType    (r));
